src/server/requirements.txt
flask
Flask-SQLAlchemy
Flask-Bcrypt
flask_cors
psycopg2-binary
python-dotenv